+++ src/main/java/io.github.kingschan1204.istock/module/maindata/services/StockService.java
@@ -6,11 +6,12 @@ import com.alibaba.fastjson.JSONObject;
 import com.mongodb.BasicDBObject;
 import com.mongodb.DBCollection;
 import com.mongodb.DBObject;
+import com.mongodb.WriteResult;
+import io.github.kingschan1204.istock.common.util.stock.StockDateUtil;
 import io.github.kingschan1204.istock.common.util.stock.StockSpider;
 import io.github.kingschan1204.istock.common.util.stock.impl.JisiluSpilder;
 import io.github.kingschan1204.istock.module.maindata.po.Stock;
 import io.github.kingschan1204.istock.module.maindata.po.StockDividend;
-import io.github.kingschan1204.istock.module.maindata.po.StockHisDividend;
 import io.github.kingschan1204.istock.module.maindata.po.StockHisRoe;
 import io.github.kingschan1204.istock.module.maindata.repository.StockHisDividendRepository;
 import io.github.kingschan1204.istock.module.maindata.repository.StockRepository;
@@ -18,11 +19,16 @@ import io.github.kingschan1204.istock.module.maindata.vo.StockVo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.mapreduce.MapReduceOptions;
+import org.springframework.data.mongodb.core.mapreduce.MapReduceResults;
 import org.springframework.data.mongodb.core.query.BasicQuery;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
+import org.springframework.data.mongodb.core.query.Update;
 import org.springframework.stereotype.Service;
+
 import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
 
@@ -46,48 +52,19 @@ public class StockService {
     @Autowired
     private JisiluSpilder jisiluSpilder;
 
+
     /**
-     * add stock code
-     *
-     * @param codes
-     * @throws Exception
+     * 查询股票
+     * @param pageindex 页码
+     * @param pagesize 显示多少条
+     * @param pcode  代码、名称
+     * @param type   市场
+     * @param pb 市净值
+     * @param dy 股息
+     * @param orderfidld 排序字段
+     * @param psort 排序方式
+     * @return
      */
-    public void addStock(String... codes) throws Exception {
-        JSONArray jsons = spider.getStockPrice(codes);
-        for (int i = 0; i < jsons.size(); i++) {
-            JSONObject json = jsons.getJSONObject(i);
-            String scode=json.getString("code");
-            JSONObject info=spider.getStockInfo(scode);
-            //"date":"2017-07-07","code":"600519","year":"2016年报","executeDate":"2017-07-01","remark":"10派67.87元(含税)","percent":1.44}
-            JSONArray dividends=spider.getHistoryDividendRate(scode);
-            JSONObject dividend;
-            String date="";
-            Double percent=0D;
-            if(null!=dividends&&dividends.size()>0){
-                for (int j = 0; j < dividends.size(); j++) {
-                    if(dividends.getJSONObject(j).getDouble("percent")>0){
-                        percent=dividends.getJSONObject(j).getDoubleValue("percent");
-                        date=dividends.getJSONObject(j).getString("date");
-                        break;
-                    }
-                }
-                //save dividend
-                List<StockHisDividend> stockHisDividendList = JSONArray.parseArray(dividends.toJSONString(),StockHisDividend.class);
-                template.remove(new Query(Criteria.where("code").is(scode)),StockHisDividend.class);
-            }
-            json.put("dividend",percent);
-            json.put("dividendDate",date);
-            json.putAll(info);
-            // his roe
-            stockHisRoeService.addStockHisRoe(scode);
-        }
-        List<Stock> list = JSON.parseArray(jsons.toJSONString(), Stock.class);
-        stockRepository.save(list);
-    }
-
-
-
-
     public String queryStock(int pageindex, int pagesize, final String pcode,final String type,String pb,String dy, String orderfidld, String psort){
         DBObject dbObject = new BasicDBObject();
         DBObject fieldObject = new BasicDBObject();
@@ -152,7 +129,7 @@ public class StockService {
 
 
     /**
-     *
+     * 得到股票历史分红信息
      * @param code
      * @return
      */
@@ -170,6 +147,11 @@ public class StockService {
     }
 
 
+    /**
+     * 得到股票历史roe
+     * @param code
+     * @return
+     */
     public List<StockHisRoe> getStockHisRoe(String code){
         Query query = new Query();
         query.addCriteria(Criteria.where("code").is(code));
@@ -255,6 +237,96 @@ public class StockService {
         result.add(pe.toString());
         result.add(reportJsons.toJSONString());
         return result;
+    }
+
+    /**
+     * 更新股票价格
+     * @param codes
+     * @param spider
+     * @throws Exception
+     */
+    public void updateStockPrice(List<String> codes,StockSpider spider) throws Exception {
+        JSONArray jsons = spider.getStockPrice(codes.toArray(new String[]{}));
+        List<Stock> stocks = JSON.parseArray(jsons.toJSONString(), Stock.class);
+        stocks.stream().forEach(stock -> {
+            template.upsert(
+                    new Query(Criteria.where("_id").is(stock.getCode())),
+                    new Update()
+                            .set("_id", stock.getCode())
+                            .set("type", stock.getType())
+                            .set("name", stock.getName())
+                            .set("price", stock.getPrice())
+                            .set("yesterdayPrice", stock.getYesterdayPrice())
+                            .set("fluctuate", stock.getFluctuate())
+                            .set("todayMax", stock.getTodayMax())
+                            .set("todayMin", stock.getTodayMin())
+                            .set("priceDate", stock.getPriceDate()),
+                    "stock"
+            );
+        });
+    }
+
+    /**
+     * 计算过去5年连续分红的平均股票股息
+     * @return
+     */
+    public void calculateFiveYearsDy() {
+        int year = StockDateUtil.getCurrentYear();
+        int fiveYearAgo = year - 5;
+        String startDate = StockDateUtil.getCurrYearLastDay(fiveYearAgo).toString();
+        String endDate = StockDateUtil.getCurrentDate();
+        Query query = new Query();
+        query.addCriteria(Criteria.where("releaseDate").gte(startDate).lte(endDate));
+        MapReduceResults<BasicDBObject> result = template.mapReduce(query, "stock_dividend",
+                "classpath:/mapreduce/5years_dy/dy5years_map.js", "classpath:/mapreduce/5years_dy/dy5years_reduce.js",
+                new MapReduceOptions().outputCollection("stock_dy_statistics"), BasicDBObject.class);
+        Iterator<BasicDBObject> iter = result.iterator();
+        while (iter.hasNext()) {
+            BasicDBObject item = iter.next();
+            String code = item.getString("_id");
+            BasicDBObject value = (BasicDBObject) item.get("value");
+            if (value.containsField("size") && value.getInt("size") > 4) {
+                double percent = Double.parseDouble(value.getString("percent"));
+                WriteResult wr = template.upsert(
+                        new Query(Criteria.where("_id").is(code)),
+                        new Update()
+                                .set("_id", code)
+                                .set("fiveYearDy", percent),
+                        "stock"
+                );
+            }
+
+        }
+    }
+
+    /**
+     * 计算上市超5年的股息roe
+     * @param startYear 开始年份
+     * @param endYear 结束年份
+     */
+    public void calculateFiveYearsRoe(int startYear,int endYear) {
+        Query query = new Query();
+        query.addCriteria(Criteria.where("year").gte(startYear).lte(endYear));
+        MapReduceResults<BasicDBObject> result = template.mapReduce(query, "stock_his_roe",
+                "classpath:/mapreduce/5years_roe/map.js", "classpath:/mapreduce/5years_roe/reduce.js",
+                new MapReduceOptions().outputCollection("stock_hisroe_statistics"), BasicDBObject.class);
+        Iterator<BasicDBObject> iter = result.iterator();
+        while (iter.hasNext()) {
+            BasicDBObject item = iter.next();
+            String code = item.getString("_id");
+            BasicDBObject value = (BasicDBObject) item.get("value");
+            if (value.containsField("size") && value.getInt("size") > 4) {
+                double percent = Double.parseDouble(value.getString("percent"));
+                WriteResult wr = template.upsert(
+                        new Query(Criteria.where("_id").is(code)),
+                        new Update()
+                                .set("_id", code)
+                                .set("fiveYearRoe", percent),
+                        "stock"
+                );
+            }
+
+        }
     }
 
 }
